Add app_instance to all metric labels
The metrics are not correctly aggregating due to multiple instances sending the
same metric/labels. We need another label to differentiate between metrics so
that Prometheus correctly aggregates them.

A change made to Prometheus also requires this label or it ends up stuck in a
loop and forever incrementing the counter.