Fix LocationSyncJob for when CSV header row is missing
The data source we use for a CSV of UK postcodes has updated their format so
that it omits the header row. I'm not sure if this has been an intentional
change, so we need to account for if the row is present or omitted.

Manually check the first row to determine if its the header and, if so, skip
it.