Migrate Hangfire database to new storage provider
Our new Postgres storage provider (compatible with Npgsql v6) is not compatible
with the structure from the old provider. Its difficult to work out what the
difference is in order to create a migration to fix it (as they are very
similar) so we can instead do a more drastic tactic of dumping the Hangfire
tables beofre the app boots, at which point Hangfire will automatically
recreate them.

We check to ensure there are no active jobs running before doing this (if there
are we will exit the app and the deploy will fail).