Use new schema name for postgres adapter
Switching to a new Hangifre Postgres adapter; the old one isn't compatible with
the new one and it needs to re-create the table. Instead of dropping the
existing table its safer to let it create a separate one; that way we can also
go in and manaully clear any jobs left in the queues.