Update school experience statuses
I accidentally used the `withdrawn` status for candidates who did not attend their placement when migrating the app to use the CandidateSchoolExperience entity. Withdrawing is a separate action.

To fix this, we have updated the `withdrawn` status name to `did not attend` (which will fix the existing requests), and added a new status for `withdrawn` (so that this can be used in the correct place).