Use new v1.2 Apply API url
The v1.2 of the Apply API is available on a different URL; this ensures its
used when the `APPLY_API_V1_2` feature is on. I would have preferred to have a
single feature switch for `APPLY_API` and an env var containing
`APPLY_API_VERSION`, however given how the conditional mapping of CRM
attributes by environment works (the attributes are only in the dev CRM at the
moment)  each version needs to have a feature switch.