Migrate non-secret environment variables into .env files
We have a number of environment variables in keyvault that are not secrets;
they are better placed in the application.

Adds additional .env files for hosted dev/test/prod environments and migrates
the non-secret environment variables from Azure keyvault.

Rename the local .env file to `.env.local` to differentiate from the hosted
development environment (which is configured with `.env.dev`).