Add GSE alert rules for job runners
We've had an issue in the past where the Delayed Job worker has hung/crashed
and jobs have backed up in GSE without us realising (until someone gets in
touch to say that they haven't received an email, for example).

Add alert rules to Prometheus for when there are a high number of jobs pending
in the queue (not being picked up) and if there are a high number of job
failures.