Upgrade to latest version of FluentValidation
The factory we were using to retrieve validators dynamically has been removed
in favour of using the service provider. There may be a cleaner way to do this,
but for now the simplest change is to inject the service provider and query the
specific validator from that (instead of the factory, which no longer exists).