Attempt to make the health check more robust
We are seeing errors on the School Experience service where the health check
times out on calling the API health check endpoint.

I'm not sure what is causing this, but I expect that the CRM is being slow to
respond as we've seen this happen in the past. To work around this we are only
going to call out to the CRM once every minute (caching the health check
responses in between). I have also changed the CRM method we call to a generic
entity name lookup for the `contact` entity; I suspect the other method we were
using may end up hitting Microsoft Authentication Library (MSAL) which could be
a reason it takes longer than expected.