Allow radius without postcode
The validation is a bit strict at the moment; if you provide a radius but no
postcode we throw an error. The front-end currently allows this scenario and it
makes sense to allow it from a users perspective as well.

Relax validation to return all results if filtering by radius without
specifying a postcode.