Update FluentValidation
The `IValidationFactory` that we use to validate models coming back from the
CRM has been deprecated and removed from the latest version of
`FluentValidation`. I haven't found a nice way of doing this validation during
reflecction with the latest version; we have to inject the whole service
provider into the models so that we can lookup the validator dynamically.

I may try refactoring this with a factory pattern in a future PR.

Also mocks out API calls in the rate limiting integration tests; I ended up
doing this as part of chasing down a test failure which was actually caused by
the validators not auto-registering with the controllers after the upgrade. The
fix was to call `AddFluentValidationAutoValidation` during startup, but I'm
leaving the mock in place as its more robust if we mock out the API (even
though the validation prevents it being it in our current test suite I can see
it being used in future tests).