Matchback on apply ID if available
When we sync candidates from the Apply service into the CRM we currently only
matchback on email address. We need to do this initially, but once the
candidate is in the CRM we could also matchback on the Apply candidate ID.

Going forward we want to matchback on the Apply candidate ID with presedence
over the email address.