Reconcile equivalent email hosts
Some email poviders allow you to use multiple hosts that direct mail to the
same mailbox. An example is `gmail.com` - if you have a GMail account you can
also send/receive email from `googlemail.com`.

The CRM team have reported that the majority of duplicate contact records
appear due to users signing up to a service with one domain and later using the
other. Instead, they want to allow matchback on any variations that are
equivilent (so point to the same mailbox/individual). This means a candidate
can sign up for, say, the mailing list with `me@gmail.com` and then later
register for an adviser using `me@googlemail.com` - the result will be a single
contact record in the CRM.

For now we are only including GMail equivalent domains as they are the main
contributor to duplicate records;  we may expand this in the future.

This functionality is behind a feature flag and will be disabled in production
until the CRM team have had a chance to test it.