Ensure CRM email takes presedence over Apply
When we matchback a candidate from the Apply service we use the Apply ID to
ensure we get the correct CRM contact even if the email is different.
Currently, we would overwrite the contact email address with the value from the
Apply candidate, however going forward we want to give the CRM contact email
presedence and retain it.