Clean up local environment variables
Historically we used `launchSettings.json` to store non-sensitive, local
environment variables. We have since introduced dotenv.net to manage our
environment variables in `.env` files.

Migrate all local environment variables into `env.local`. Update the `Env`
class to default the `CloudFoundryEnvironmentName` to `local` if the
`VCAP_APPLICATION` environemnt variable is not set (this is a bit of a
chicken/egg problem as the environment name is determined by the
`VCAP_APPLICATION` environment variable, which isn't available until dotenv
loads it).

The `ASPNETCORE_ENVIRONMENT` will remain in `launchSettings.json` as it needs
to be set prior to the application booting.