Queue GSE sign up when CRM is offline
By default we create GSE candidates in-band as the service needs the created
`contact.id` from the CRM in order to match up the GSE candidate with the CRM
contact.

When the CRM integration is paused (this happens for maintenance occasionally)
we don't want to prevent GSE sign ups, so we fallback to generating a
`contact.id` ahead of persisting it in the CRM and we queue the upsert
operation.

We had an incident a number of weeks ago where the CRM was offline unexpectedly
for a period of time without being 'paused' (which the CRM team do manually
prior to scheduled maintenance). During this time GSE sign ups were failing as
the CRM could not be reached. Instead, we want to detect when the CRM is having
issues and is _not_ in a paused state so that we can fallback to generating the
`contact.id` ahead of time and queueing the upsert operation so that it can run
when the CRM recovers.