Allow zero callback booking quotas to be returned
Currently we have fallback logic so that if the CRM is offline or 0 callback
slots are available we are still going to return a default set of slots that
will allow candidates to continue their sign up.

We recently had an unexpected influx of TTA sign ups that were being offered
callback booking slots and found we had no way to stem the flow of bookings (if
we mark all slots as available the default slots would then be offered). We
currently offer 14 days in advance, so the liklihood of all slots being booked
are slim, but we want to account for the scenario in case it does happen in the
future.

Update the callback service to only return fallback slots when the CRM is
offline. If the CRM is indicating there are no available slots we want to
surface this on the websites and so we will now return 0 slots available.