Make phone call/telephone optional for equivalent degrees
Currently, if a candidate has an equivalent degree we require them to schedule
a callback with an adviser. Going forward, we are allowing these candidates to
sign up without a callback in the event that there are no available callback
booking slots available.

Trying to enforce the validation rules for when there are no callback slots
available would be error prone, given that the availability dynamically
updates. Instead, we are going to relax the validation in the API and ensure
its enforced in the client services.