Fix Docker build restore issue
I'm not sure what changed, but the `dotnet publish --no-restore` started
failing due to not locating a package. By doing the restore as part of the
publish instead of in an earlier step it seems to fix it. There may be a slight
performance hit doing it this way, but the docker build is reasonably fast
anyway.