Add individual Country/TeachingSubject models
We currently use a generic `LookupItem` model for both countries and teaching
subjects. Unlike `PickListItem` these entites in the CRM can have additional
fields assigned to them. The countries will now be returned with a country
code, so we need the ability to add fields to our models.

I originall started modelling the `Country` as a subclass of `LookupItem` but
it quickly unravelled and I found migrating the current `LookupItem`s would be
difficult this way. Instead, we're adding separate `Country` and
`TeachingSubject` models. These will retain the `Id` and `Value` attributes of
`LookupItem` so the client applications can continue to function without
changes.