Always retain existing candidate email
As we are now matching on equivalent email domains and Apply ID we can run into
the situation where the candidate supplies a different email address to the one
we have against their candidate record. In this scenario we want to always
retain the existing candidate email as TPUK use this to allocate candidates to
advisers; if we change it we could disrupt that process.