Update adviser assignment logic
Currently, if a candidate has a degree or is returning to teaching we will put
them in the queue to be assigned to an adviser. Other candidates are taken on a
journey to schedule a callback on the main TTA service.

If an international candidate signs up from Apply we do not offer the callback
service and they aren't put in the queue for an adviser, so they will fall
between the cracks. Going forward we want to queue these candidates for an
adviser. To that end, the logic around this can be simplified to 'if a
candidate has not booked a callback, put them in the queue for an adviser'.