Disable Hanfire Prometheus metrics exporter
We recently performed an Apply backfill in production which has resulted in
many more records being in the database than usual. The logs indicate the
Hangfire exporter is perhaps running an inefficient query and flooding the
database with requests it can't keep up with.

Disable Hangfire Prometheus exporter in all environments.