Sync events over the last 12 months
We currently only sync events going back 4 months and we are finding that users
are coming across old links to past events from external sources, resulting in
a 404 not found error. Instead, if the event did exist but has since passed we
would be better to serve a 410 gone with a better error message for the user.
The GiT website does this already, but only as far back as the API is syncing;
increasing this to 12 months to avoid 404s on recently(ish) past events.